sdsrange: change start/end args to ptrdiff_t
An sds string can hold a much bigger value than an int can hold, but
since size_t is unsigned it can’t be used for this function. Using
ptrdiff_t limits the function to work on roughly 1 Exabyte long strings,
but if you need to work with longer ones you probably will have your
custom string library already.

+++ src/sds.c
@@ -27,6 +27,7 @@
 
 #include <assert.h>
 #include <ctype.h>
+#include <stddef.h>
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
@@ -388,7 +389,7 @@ void sdstrim(sds s, const char *cset) {
  * s = sdsnew("Hello World");
  * sdsrange(s,1,-1); => "ello World"
  */
-void sdsrange(sds s, int start, int end) {
+void sdsrange(sds s, ptrdiff_t start, ptrdiff_t end) {
 	struct sdshdr *sh = sdsheader(s);
 	size_t newlen, len = sdslen(s);
